Refactoring Redis ClientBase out of security groups
To make room for other uses of the Redis cache, it
is necessary to refactor the core components of a
ClientBase out of the existing Security Groups
Redis client. Afterwards I created a child
SecurityGroupsClient to implement the Security
Groups specific features.
